+++ sdk/keyvault/keyvault-keys/src/api/keyVaultContext.ts
@@ -42,7 +42,7 @@ export function createKeyVault(
   };
   const clientContext = getClient(endpointUrl, credential, updatedOptions);
   clientContext.pipeline.removePolicy({ name: "ApiVersionPolicy" });
-  const apiVersion = options.apiVersion ?? "2026-03-01-preview";
+  const apiVersion = options.apiVersion ?? "2025-07-01";
   clientContext.pipeline.addPolicy({
     name: "ClientApiVersionPolicy",
     sendRequest: (req, next) => {